requirements.txt
streamlit>=1.28.0
py3Dmol>=2.0.4
biopython>=1.81